refactored and moved IT test files
